% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_methods.R
\name{extract_class}
\alias{extract_class}
\title{Extract R6 Class}
\usage{
extract_class(content, start_pos = 1)
}
\arguments{
\item{content}{Character, content of a file or a string}

\item{start_pos}{Integer, row position of cursor. Serves as starting point to search
for class definition}
}
\value{
A list with fields
\itemize{
  \item{class_content}{Character, extracted class definition}
  \item{start}{Integer, start position of class definition within `content`}
  \item{end}{Integer, end position of class definition within `content`}
}
}
\description{
Extract R6 Class
}
