% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closing.R
\name{find_closing}
\alias{find_closing}
\title{Find Closing}
\usage{
find_closing(text, opening = "\\\\(", closing = "\\\\)")
}
\arguments{
\item{text}{Character, text to search}

\item{opening}{Opening character}

\item{closing}{Closing character}
}
\value{
Integer
}
\description{
Find position of closing character to first encountered opening character
}
