% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{get_cursor_pos}
\alias{get_cursor_pos}
\title{Get Cursor Position From Active Document}
\usage{
get_cursor_pos(x)
}
\arguments{
\item{x}{Document Context}
}
\value{
Integer, position of cursor in text
}
\description{
Get Cursor Position From Active Document
}
