% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_methods.R
\name{insert_methods}
\alias{insert_methods}
\title{Insert Methods}
\usage{
insert_methods(content, cursor_row)
}
\arguments{
\item{content}{Character, content of the file or a string}

\item{cursor_row}{Integer, row position of cursor}
}
\value{
Character, modified \code{content} with injected methods
}
\description{
Insert Methods
}
