% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insert_methods_addin}
\alias{insert_methods_addin}
\title{An addin for inserting methods straigth into the source file}
\usage{
insert_methods_addin()
}
\description{
An addin for inserting methods straigth into the source file
}
