% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{iter}
\alias{iter}
\alias{is.iterable}
\title{Make R6 Instance Iterable}
\usage{
iter(x)

is.iterable(x)
}
\arguments{
\item{x}{\code{R6} object.}
}
\description{
This function uses \code{coro} to create a generator.
Using the returned object (a function), you can call this function to
create an iterator. The iterator can be called repeatedly
to produce batches of data. When all of the batches have been produced,
the function will return the symbol \code{exhausted} for each call thereafter.
}
\details{
A \code{coro} iterator is compatible with \code{for} loops via the
\code{coro::loop()} function. \code{coro::collect()} can be used to collect the.
next \code{n} set of batches. Leaving \code{n} blank will drain all of the remaining
batches (assuming \code{length(x)} is not infinite). See \code{coro}
documentation for more details.
}
