% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{make_gadget}
\alias{make_gadget}
\title{Make Gadget}
\usage{
make_gadget(title, title_bar, done_fun)
}
\arguments{
\item{title}{Character, title of gadget window}

\item{title_bar}{Character, gadget title bar}

\item{done_fun}{Function to be used after clicking `Done` button}
}
\description{
Create gadget for generating R6 methods. Action after clicking `Done` button is defined
by `done_fun`.
}
