% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_methods.R
\name{make_method_str}
\alias{make_method_str}
\alias{make_getter_method_str}
\alias{make_setter_method_str}
\title{Make method string}
\usage{
make_getter_method_str(field, is_public = TRUE, add_roxygen = TRUE)

make_setter_method_str(field, is_public = TRUE, add_roxygen = TRUE)
}
\arguments{
\item{field}{Character name of class field}

\item{is_public}{Logical, whether the field is in public list}

\item{add_roxygen}{Logical, whether to add roxygen description of method}
}
\value{
Character containing method definition
}
\description{
Make method string
}
