% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_getter.R
\name{make_methods}
\alias{make_methods}
\title{Make methods}
\usage{
make_methods(
  r6,
  field = c("both", "public", "private", names(r6$public_fields),
    names(r6$private_fields)),
  method = c("both", "get", "set"),
  add_roxygen = TRUE
)
}
\arguments{
\item{r6}{R6 class for which to create methods}

\item{field}{Character, fields for which to create method. May be "both",
"public", "private" or name of class field. Multiple values allowed.}

\item{method}{Character, methods to create. One of "both", "get", "set"}

\item{add_roxygen}{Logical, whether to add roxygen description of method}
}
\value{
Character containing generated methods to put into class definition
}
\description{
Make methods
}
