% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{make_methods_addin_gadget}
\alias{make_methods_addin_gadget}
\title{Make methods addin gadget}
\usage{
make_methods_addin_gadget()
}
\value{
No return value, called for side effects
}
\description{
Make methods addin gadget
}
