% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_r6.R
\name{ReactiveR6}
\alias{ReactiveR6}
\title{ReactiveR6}
\description{
This class allows you to make your R6 class reactive by inheriting from \code{ReactiveR6}
By calling \code{private$invalidate()} in a method you can invalidate the class in a controlled way, i.e.
only when specific methods are called. See example how to use it.

Inspired by \url{https://community.rstudio.com/t/good-way-to-create-a-reactive-aware-r6-class/84890}
}
\examples{
\dontrun{
if (interactive()) {
library(shiny)
library(r6methods)

Counter <-  R6::R6Class(
  "Counter",
  inherit = ReactiveR6,
  public = list(
    increment = function() {
      private$counter <- private$counter + 1
      private$invalidate()
    },
    decrement = function() {
      private$counter <- private$counter - 1
      private$invalidate()
    },
    silent_increment = function() {
      private$counter <- private$counter + 1
    },
    get_counter = function() {
      private$counter
    }
  ),
  private = list(
    counter = 0
  )
)

counter <- Counter$new()$reactive()

shinyApp(
  fluidPage(
    actionButton("increment", "Increment"),
    actionButton("decrement", "Decrement"),
    actionButton("silent_increment", "Silent increment"),
    textOutput("value")
  ),
  function(input, output, session) {
    observeEvent(input$increment, {
      counter()$increment()
    })

    observeEvent(input$decrement, {
      counter()$decrement()
    })

    observeEvent(input$silent_increment, {
      counter()$silent_increment()
    })

    output$value <- renderText({
      counter()$get_counter()
    })
  }
)
}
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-reactive}{\code{ReactiveR6$reactive()}}
\item \href{#method-clone}{\code{ReactiveR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reactive"></a>}}
\if{latex}{\out{\hypertarget{method-reactive}{}}}
\subsection{Method \code{reactive()}}{
Call this method to make object instance reactive
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveR6$reactive()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
