% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_methods.R
\name{source_class}
\alias{source_class}
\title{Source class}
\usage{
source_class(txt)
}
\arguments{
\item{txt}{Character, text containing class definition}
}
\value{
R6 class
}
\description{
Sources R6 class from text, prepends namespace to `R6Class` in order to not
require `R6` to be loaded.
}
