% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_methods.R
\name{insert_methods}
\alias{insert_methods}
\title{Insert Methods}
\usage{
insert_methods(
  content,
  start_pos = 1,
  field = c("all", "public", "private"),
  method = c("both", "get", "set"),
  add_roxygen = TRUE
)
}
\arguments{
\item{content}{Character, content of the file or a string}

\item{start_pos}{Integer, position of cursor within `content`. Number of characters
before the cursor.}

\item{field}{Character, fields for which to create method. May be "all",
"public", "private" or name of class field. Multiple values allowed.}

\item{method}{Character, methods to create. One of "both", "get", "set"}

\item{add_roxygen}{Logical, whether to add roxygen description of method}
}
\value{
Character, modified \code{content} with injected methods
}
\description{
Insert Methods
}
