% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_getter.R
\name{make_getter_method_str}
\alias{make_getter_method_str}
\alias{make_setter_method_str}
\title{Make method string}
\usage{
make_getter_method_str(field, add_roxygen = TRUE)

make_setter_method_str(field, add_roxygen = TRUE)
}
\arguments{
\item{field}{Character name of class field}

\item{add_roxygen}{Logical, whether to add roxygen description of method}
}
\value{
Character
}
\description{
Make method string
}
